removed time_slot_helper and combined the functions inside it to scheduling service
